split eltex detection into utility class
